Move precondition checks from DefaultKernelUtils to Preconditions class
