+++ kernel/kernel-defaults/src/main/java/io/delta/kernel/defaults/internal/DefaultKernelUtils.java
@@ -72,51 +72,6 @@ public class DefaultKernelUtils {
         return null;
     }
 
-    // TODO: Move these precondition checks into a separate utility class.
-
-    /**
-     * Precondition-style validation that throws {@link IllegalArgumentException}.
-     *
-     * @param isValid {@code true} if valid, {@code false} if an exception should be thrown
-     * @throws IllegalArgumentException if {@code isValid} is false
-     */
-    public static void checkArgument(boolean isValid)
-        throws IllegalArgumentException {
-        if (!isValid) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    /**
-     * Precondition-style validation that throws {@link IllegalArgumentException}.
-     *
-     * @param isValid {@code true} if valid, {@code false} if an exception should be thrown
-     * @param message A String message for the exception.
-     * @throws IllegalArgumentException if {@code isValid} is false
-     */
-    public static void checkArgument(boolean isValid, String message)
-        throws IllegalArgumentException {
-        if (!isValid) {
-            throw new IllegalArgumentException(message);
-        }
-    }
-
-    /**
-     * Precondition-style validation that throws {@link IllegalArgumentException}.
-     *
-     * @param isValid {@code true} if valid, {@code false} if an exception should be thrown
-     * @param message A String message for the exception.
-     * @param args    Objects used to fill in {@code %s} placeholders in the message
-     * @throws IllegalArgumentException if {@code isValid} is false
-     */
-    public static void checkArgument(boolean isValid, String message, Object... args)
-        throws IllegalArgumentException {
-        if (!isValid) {
-            throw new IllegalArgumentException(
-                String.format(String.valueOf(message), args));
-        }
-    }
-
     private static List<Type> pruneFields(GroupType type, StructType deltaDataType) {
         // prune fields including nested pruning like in pruneSchema
         return deltaDataType.fields().stream()
