+++ kernel/kernel-defaults/src/main/java/io/delta/kernel/defaults/internal/data/vector/DefaultLongVector.java
@@ -22,7 +22,7 @@ import io.delta.kernel.types.DataType;
 import io.delta.kernel.types.LongType;
 import io.delta.kernel.types.TimestampType;
 
-import static io.delta.kernel.defaults.internal.DefaultKernelUtils.checkArgument;
+import static io.delta.kernel.defaults.internal.Preconditions.checkArgument;
 
 /**
  * {@link io.delta.kernel.data.ColumnVector} implementation for long or timestamp type data.
