+++ kernel/kernel-defaults/src/main/java/io/delta/kernel/defaults/internal/parquet/DecimalConverters.java
@@ -34,7 +34,7 @@ import io.delta.kernel.types.DecimalType;
 
 import io.delta.kernel.defaults.internal.data.vector.DefaultDecimalVector;
 import io.delta.kernel.defaults.internal.parquet.ParquetConverters.BasePrimitiveColumnConverter;
-import static io.delta.kernel.defaults.internal.DefaultKernelUtils.checkArgument;
+import static io.delta.kernel.defaults.internal.Preconditions.checkArgument;
 
 public class DecimalConverters {
