+++ kernel/kernel-defaults/src/main/java/io/delta/kernel/defaults/internal/parquet/RepeatedValueConverter.java
@@ -23,7 +23,7 @@ import org.apache.parquet.io.api.GroupConverter;
 import io.delta.kernel.data.ColumnVector;
 
 import io.delta.kernel.defaults.internal.parquet.ParquetConverters.BaseConverter;
-import static io.delta.kernel.defaults.internal.DefaultKernelUtils.checkArgument;
+import static io.delta.kernel.defaults.internal.Preconditions.checkArgument;
 import static io.delta.kernel.defaults.internal.parquet.ParquetConverters.initNullabilityVector;
 import static io.delta.kernel.defaults.internal.parquet.ParquetConverters.setNullabilityToTrue;
